python-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.6.0
pydantic==2.9.2
python-dotenv==1.0.1
python-multipart==0.0.9
cloudinary==1.41.0
pywebpush==2.0.0
cryptography==43.0.1
bcrypt==4.1.3
PyJWT==2.9.0
orjson==3.10.7
redis[hiredis]==5.0.4
httpx==0.27.2
pytest==8.3.2
pytest-asyncio==0.23.7
mongomock-motor==0.0.17
